[Synthetics] Trim TS project refs via type-only dep extraction
Synthetics referenced three heavy `-plugin` TS projects purely for
type-only usages, dragging their whole subtrees into its type-build.
Replace them with lightweight sources so those subtrees drop out:

- Relocate `DataStream` (and related data-stream types + `IndexMode`)
  into `@kbn/index-management-shared-types`; `@kbn/index-management-plugin`
  common re-exports them for back-compat. Synthetics now imports
  `DataStream` from the shared-types package instead of the plugin,
  removing the streams/dataset-quality/upgrade-assistant subtree.
- Replace the type-only `SLOPublicStart` import with a minimal local
  contract whose `initialValues` stays typed via `@kbn/slo-schema`
  (already referenced), dropping `@kbn/slo-plugin`.
- Replace the type-only `LicenseManagementUIPluginSetup` import (only
  `.enabled` is used) with a minimal local type, dropping
  `@kbn/license-management-plugin`.

Runtime plugin dependencies in kibana.jsonc are unchanged; this only
affects TS project references. No behavior change.
