+++ x-pack/platform/packages/shared/index-management/index_management_shared_types/index.ts
@@ -10,3 +10,5 @@ export * from './src/index_templates';
 export * from './src/home_sections';
 export type * from './src/services';
 export * from './src/constants';
+export * from './src/index_modes';
+export type * from './src/data_streams';
