+++ x-pack/platform/packages/shared/index-management/index_management_shared_types/src/index_modes.ts
@@ -1,0 +1,22 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const STANDARD_INDEX_MODE = 'standard';
+export const LOGSDB_INDEX_MODE = 'logsdb';
+export const TIME_SERIES_MODE = 'time_series';
+export const LOOKUP_INDEX_MODE = 'lookup';
+export const VECTOR_DB_INDEX_MODE = 'vectordb_document';
+
+export const IndexMode = {
+  standard: STANDARD_INDEX_MODE,
+  logsdb: LOGSDB_INDEX_MODE,
+  time_series: TIME_SERIES_MODE,
+  lookup: LOOKUP_INDEX_MODE,
+  vectordb_document: VECTOR_DB_INDEX_MODE,
+} as const;
+
+export type IndexMode = (typeof IndexMode)[keyof typeof IndexMode];
