+++ x-pack/solutions/observability/plugins/synthetics/moon.yml
@@ -88,8 +88,7 @@ dependsOn:
   - '@kbn/securitysolution-io-ts-types'
   - '@kbn/shared-ux-link-redirect-app'
   - '@kbn/serverless'
-  - '@kbn/index-management-plugin'
-  - '@kbn/license-management-plugin'
+  - '@kbn/index-management-shared-types'
   - '@kbn/react-kibana-mount'
   - '@kbn/react-kibana-context-render'
   - '@kbn/react-kibana-context-theme'
@@ -104,7 +103,6 @@ dependsOn:
   - '@kbn/slo-schema'
   - '@kbn/alerting-types'
   - '@kbn/babel-register'
-  - '@kbn/slo-plugin'
   - '@kbn/ebt-tools'
   - '@kbn/alerting-types'
   - '@kbn/core-chrome-browser'
