+++ x-pack/solutions/observability/plugins/synthetics/public/apps/synthetics/components/settings/hooks/use_get_data_stream_statuses.test.ts
@@ -10,7 +10,7 @@ import { renderHook } from '@testing-library/react';
 import { FETCH_STATUS, useFetcher } from '@kbn/observability-shared-plugin/public';
 import { useGetDataStreamStatuses } from './use_get_data_stream_statuses';
 import { policyLabels } from '../data_retention/policy_labels';
-import type { DataStream } from '@kbn/index-management-plugin/common';
+import type { DataStream } from '@kbn/index-management-shared-types';
 
 jest.mock('@kbn/observability-shared-plugin/public', () => ({
   ...jest.requireActual('@kbn/observability-shared-plugin/public'),
