+++ x-pack/solutions/observability/plugins/synthetics/tsconfig.json
@@ -85,8 +85,7 @@
     "@kbn/securitysolution-io-ts-types",
     "@kbn/shared-ux-link-redirect-app",
     "@kbn/serverless",
-    "@kbn/index-management-plugin",
-    "@kbn/license-management-plugin",
+    "@kbn/index-management-shared-types",
     "@kbn/react-kibana-mount",
     "@kbn/react-kibana-context-render",
     "@kbn/react-kibana-context-theme",
@@ -101,7 +100,6 @@
     "@kbn/slo-schema",
     "@kbn/alerting-types",
     "@kbn/babel-register",
-    "@kbn/slo-plugin",
     "@kbn/ebt-tools",
     "@kbn/alerting-types",
     "@kbn/core-chrome-browser",
